+++ credits/views.py
@@ -1,0 +1,6 @@
+from django.http import HttpResponse
+
+
+def placeholder(_request):
+    return HttpResponse("credits ok")
+